+++ base/documentation.py
@@ -6,6 +6,7 @@ import inspect
 import base
 import attrib
 import xml.etree.ElementTree
+import textwrap
 
 
 # CHANGELOG
@@ -14,6 +15,20 @@ base.VERSION.added("1.5.0", "`base.documentation` methods for documenting compon
 base.VERSION.added("1.5.1", "`base.documentation` methods for documenting scenarios")
 base.VERSION.added("1.5.3", "`base.documentation.write_changelog()` no longer escapes underscores")
 base.VERSION.added("1.5.7", "`base.documentation.document_component()` documentation of `Equals` attribute")
+base.VERSION.changed(
+    "1.5.8",
+    """`base.documentation.document_component()` can now handle sample configurations with component names differing 
+    from object name"""
+)
+base.VERSION.changed(
+    "1.5.8",
+    """`base.documentation.document_component()`: long lines in XML samples are now wrapped to ensure 120 character 
+    width"""
+)
+base.VERSION.added(
+    "1.5.8",
+    "`base.documentation.document_component()` documentation of data_type attribute hint and default attribute"
+)
 
 
 def write_changelog(name, version_history, file_path):
@@ -98,15 +113,18 @@ def document_stores(stores_module, file_path):
     return
 
 
-def document_component(component, file_path, sample_configuration):
+def document_component(component, file_path, sample_configuration, sample_component_name=None):
     """
     :param component: The component to document.
     :param file_path: The path of file where the readme is written to.
     :param sample_configuration: The path to a configuration file where the component is configured.
+    :param sample_component_name: The name of the component in the sample configuration if it does not equal the
+    component name.
     :return: Nothing.
     """
     module_doc = "" if component.module.doc_file is None else "(see `" + component.module.doc_file + "` for details)"
     configuration_xml = xml.etree.ElementTree.parse(sample_configuration)
+    component_name = component.name if sample_component_name is None else sample_component_name
     with open(file_path, "w", encoding="utf-8") as f:
         f.write("""## Table of Contents
 * [About the project](#about-the-project)
@@ -166,9 +184,14 @@ The following gives a sample configuration of the `{component_name}` component. 
             component_name=component.name,
             component_module=component.__module__,
             component_class=type(component).__name__,
-            sample_configuration=inspect.cleandoc(
-                xml.etree.ElementTree.tostring(
-                    configuration_xml.getroot().find("Composition").find(component.name)).decode("utf-8"))
+            sample_configuration="\n".join(
+                textwrap.wrap(
+                    inspect.cleandoc(
+                        xml.etree.ElementTree.tostring(
+                            configuration_xml.getroot().find("Composition").find(component_name)).decode("utf-8")),
+                    120,
+                    replace_whitespace=False
+                ))
         ))
         for component_input in component.inputs:
             f.write("#### {}\n".format(component_input.name))
@@ -211,6 +234,8 @@ The following gives a sample configuration of the `{component_name}` component. 
                         f.write("Dimension {} spans {}.\n".format(i + 1, value[i]))
                 elif attribute == "chunks":
                     f.write("Chunking of the array is {}.\n".format(value))
+                elif attribute == "data_type":
+                    f.write("Individual array elements have a type of `{}`.\n".format(value.__name__))
                 else:
                     raise ValueError("Unsupported default attribute: " + attribute)
             for attribute, value in component_output.default_attributes.items():
@@ -223,6 +248,8 @@ The following gives a sample configuration of the `{component_name}` component. 
                         f.write("Values have no physical unit.\n")
                     else:
                         f.write("The physical unit of the values is `{}`.\n".format(value))
+                elif attribute == "default":
+                    f.write("The default value of the output is `{}`.\n".format(value))
                 else:
                     raise ValueError("Unsupported default attribute: " + attribute)
         f.write("\n\n## Roadmap\n")